Proyecto Pollino-Gabriel fusionado

+++ Datos/Infrastructure/Actividad.cs
@@ -24,7 +24,7 @@ namespace Datos.Infrastructure
         public string Nombre { get; set; }
         public string Descripcion { get; set; }
         public string DNI_Monitor { get; set; }
-        public Nullable<byte> Media_Valoracion { get; set; }
+        public Nullable<decimal> Media_Valoracion { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
         public virtual ICollection<Usuario_Actividad> Usuario_Actividad { get; set; }
