+++ Negocio/EntitiesDTO/ActividadDTO.cs
@@ -14,7 +14,7 @@ namespace Negocio.EntitiesDTO
         public string Descripcion { get; set; }
         public string DNI_Monitor { get; set; }
 
-        public byte MediaValoracion { get; set; }
+        public double MediaValoracion { get; set; }
 
         public ActividadDTO()
         {
